# Avoid reversing the slice