--- a/balloon.go
+++ b/balloon.go
@@ -130,8 +130,9 @@ func (b *Instance) Mix(h hash.Hash, salt []byte, sCost, tCost uint64) {
 				otherBytes := h.Sum(nil)
 
 				// math/big usually uses big-endian, with the exception of
-				// SetBits. We can unsafely convert otherBytes directly into
-				// a big.Word slice to bypass the need to reverse otherBytes.
+				// Int.SetBits. We can unsafely convert otherBytes directly
+				// into a big.Word slice to bypass the need to reverse
+				// otherBytes.
 				header := *(*reflect.SliceHeader)(unsafe.Pointer(&otherBytes))
 				header.Len /= SizeOfWord
 				header.Cap /= SizeOfWord