This PR aims to remove the need to reverse the bytes slice when using math/big. I also made some of the variable/type naming more consistent (Go uses camel-case) and fixed a few typos.

I saw that there was an issue with math/big using primarily big-endian in its package. Oddly, I came across this function https://pkg.go.dev/math/big#Int.SetBits that uses little-endian.

It's possible (in a rather hacky way) to use the unsafe package to type cast the bytes slice into a big.Word slice and use SetBits without the need to reverse a slice.

Here are the given benchmark tests, showing a slight improvement.

With reverse
```
BenchmarkBalloon
BenchmarkBalloon-4          6135           1736703 ns/op
BenchmarkBalloonM
BenchmarkBalloonM-4         4665           2638889 ns/op
```

Without reverse
```
BenchmarkBalloon
BenchmarkBalloon-4          6474           1692983 ns/op
BenchmarkBalloonM
BenchmarkBalloonM-4         4732           2511378 ns/op
```

I think more can be squeezed out of this hashing implementation, but it'll need more looking into.