@daxpedda 2022-06-09
I apparently wasn't subscribed to this repo, so I didn't see your PR. I don't really feel qualified to review this, my knowledge of Go is not limited, it's zero. So whenever @nogoegst get's around reviewing the original PR, he can take a look here too.

Thanks for looking into it though!