@darenliang 2022-06-09
It is all good,

I was learning about balloon hashing and came across your PR. After looking at the password-hashes crate, I'm jealous of Rust devs. I really hope someone can create a Go library with a collection of modern hashing algs.